--- pages/country/data.js
@@ -1,23 +1,0 @@
-const timeLineEducationProps = [
-  {
-    title: 'Materno infantil',
-    color: '#fb8080',
-    width: '21.5%',
-  },
-  {
-    title: 'Preescolar',
-    color: '#bc6060',
-    width: '10.6%',
-  },
-  {
-    title: ' Educación general básica',
-    color: '#7ab239',
-    width: '50%',
-  },
-  {
-    title: 'Educación diversificada',
-    color: '#0071bc',
-    width: '15.1%',
-  },
-];
-export default timeLineEducationProps;
